Refactoring : imageUploader
비즈니스로직 hook으로 분리
